# contracts/feedback/openapi.yaml
openapi: 3.1.0
...
info:
  title: Feedback Service API
  version: 1.0.0-alpha.1
  description: |
    ホスト業務システムから独立したフィードバック基盤の公開契約。
    Web GIS の project、GIS、業務 CRUD、タイル API を含めない。
servers:
  - url: /feedback/v1
security:
  - bearerAuth: []
...
paths:
...
  /memberships:
    get:
      tags: [administration]
      operationId: listFeedbackWorkspaceMemberships
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
      responses:
        "200":
          description: workspace membership 一覧
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FeedbackWorkspaceMember" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [administration]
      operationId: createFeedbackWorkspaceMembership
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackMembershipCreateRequest" }
      responses:
        "201":
          description: 追加したmembership
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackWorkspaceMember" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "416": { $ref: "#/components/responses/RangeNotSatisfiable" }
        "409": { $ref: "#/components/responses/Conflict" }
  /memberships/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ApplicationKeyQuery"
      - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
      - $ref: "#/components/parameters/IfMatch"
    patch:
      tags: [administration]
      operationId: patchFeedbackWorkspaceMembership
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/FeedbackMembershipPatchRequest" }
      responses:
        "200":
          description: 更新したmembership
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackWorkspaceMember" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
    delete:
      tags: [administration]
      operationId: deleteFeedbackWorkspaceMembership
      responses:
        "204": { description: membershipを削除 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        直接OIDC JWTまたは契約済みtoken exchange JWT。直接OIDC JWTは検証対象のissuer、audience、
        時刻、設定済みsubject claimに加え、`feedback_permissions`を1件以上のFeedbackPermissionからなる
        文字列配列として必須とする。直接OIDCのpermissionは全resourceでDB membershipと交差する。
        token exchange JWTのresource scopeとclaimはtoken exchange公開契約に従う。
  parameters:
...
    ApplicationKeyQuery:
      name: applicationKey
      in: query
      required: true
      schema: { type: string, pattern: "^[a-z][a-z0-9-]{0,62}$" }
...
    ExternalWorkspaceKeyQuery:
      name: externalWorkspaceKey
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, minLength: 3, maxLength: 200 }
...
  headers:
    ETag:
      description: resource version。更新時は If-Match へそのまま指定する。
      schema: { type: string }
  responses:
    BadRequest:
      description: 契約または manifest に適合しない入力
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Unauthorized:
      description: 未認証または token 無効
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Forbidden:
      description: feedback permission 不足
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    NotFound:
      description: resource 不在または存在秘匿
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Conflict:
      description: open session 競合などの一意制約違反
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    PreconditionFailed:
      description: ETag が古い
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
    RangeNotSatisfiable:
      description: evidence size の範囲外または複数 byte range
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
  schemas:
...
    FeedbackProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string, pattern: "^[a-z0-9_.-]+$" }
        requestId: { type: string, minLength: 1, maxLength: 200 }
    FeedbackPermission:
      enum: [feedback.read, feedback.comment, feedback.manage, feedback.admin]
...
    FeedbackWorkspaceMember:
      type: object
      additionalProperties: false
      required: [userId, issuer, subject, permissions, version]
      properties:
        userId: { type: string, format: uuid }
        issuer: { type: string, minLength: 1, maxLength: 1000 }
        subject: { type: string, minLength: 1, maxLength: 200 }
        email: { type: [string, "null"], maxLength: 320 }
        displayName: { type: [string, "null"], maxLength: 200 }
        permissions:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/FeedbackPermission" }
        version: { type: integer, minimum: 1 }
    FeedbackMembershipCreateRequest:
      type: object
      additionalProperties: false
      required: [issuer, subject, permissions]
      properties:
        issuer: { type: string, minLength: 1, maxLength: 1000 }
        subject: { type: string, minLength: 1, maxLength: 200 }
        permissions:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/FeedbackPermission" }
    FeedbackMembershipPatchRequest:
      type: object
      additionalProperties: false
      required: [permissions]
      properties:
        permissions:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/FeedbackPermission" }
...